feat!: refresh the accepted protocol to OpenCode 2.0.5
Move the accepted OpenAPI snapshot, the Restore patch, and the source watch to
upstream 2.0.5 (79169fe966d58fb0e0a5e41133716184a0ab4ca6) under receipt, and
regenerate the SDK against it with no compatibility layer.

Protocol and generator
- Operation identities are prefixless; the identity wall refuses any
  `server.<name>.<endpoint>` id, not one literal, and repair rows for such ids
  bind again. `form.list` is `ListFormsAsync` / `Forms`; a `schemaNames` row
  that names a stabilize duplicate the collapse folds is refused as redundant.
- 131 of 134 operations are callable: 129 generated plus the two WebSocket
  transports; `config.get`, `experimental.migration.v1.status`, and `fs.read`
  stay declined with their reasons.

Runtime
- `GetHealthAsync` becomes `Server.GetStatusAsync`; the discovery probe is the
  status probe and carries upstream's `compatible` answer for an authenticated
  404, which discovery drops. Location targeting is directory-only; the
  experimental operations live on the flat `ExperimentalClient`.

Tests and docs
- Contract tests assert the envelope `Location.Directory`; plugin inventory
  tests assert on both branches with the ADR-0022 rationale restored.
- CHANGELOG lists the breaking changes from the API baseline; README, ROADMAP,
  ADR-0005/0008/0011/0019/0021/0025, the guides, and CONTEXT.md describe the
  pinned state.

BREAKING CHANGE: see the CHANGELOG [Unreleased] section — renamed clients and
operations, the status-based health surface, directory-only locations, and
removed 1.x-era members.

+++ src/OpenCode.Sdk/Models/SessionViewPostRequest.cs
@@ -12,5 +12,5 @@ public sealed record SessionViewPostRequest
     /// Gets the idle value.
     /// </summary>
     [JsonPropertyName("idle")]
-    public required long Idle { get; init; }
+    public required double Idle { get; init; }
 }
